moved the containers folder up out of base
