+++ docs/index.html
@@ -35,5 +35,5 @@
   </head>
   <body>
     <div id="app"></div>
-  <script type="text/javascript" src="build/bundle.c6314766.js"></script></body>
+  <script type="text/javascript" src="build/bundle.125548c4.js"></script></body>
 </html>
