+++ package.json
@@ -16,7 +16,7 @@
     "test": "jest"
   },
   "lint-staged": {
-    "*.js": [
+    "src/**/*.js": [
       "prettier --single-quote --write",
       "eslint --fix",
       "git add"
@@ -69,8 +69,8 @@
     "babel-plugin-lodash": "^3.2.11",
     "babel-runtime": "^6.20.0",
     "classnames": "^2.2.5",
-    "focus-trap-react": "^3.0.3",
     "date-fns": "^1.28.3",
+    "focus-trap-react": "^3.0.3",
     "lodash": "^4.17.4",
     "react-overlays": "^0.7.0",
     "react-smooth-collapse": "^1.2.0",
