--- src/components/base/containers/Containers.md
@@ -1,1 +1,0 @@
-Container components handle wrapping child components with the appropriate margins and padding.
