--- src/components/base/containers/alert/alert-variations.js
@@ -1,29 +1,0 @@
-import { colors, backgroundColors } from '../../../theme';
-
-export const alertVariations = {
-  success: {
-    color: backgroundColors.success,
-    borderColor: colors.success,
-    foregroundColor: colors.success
-  },
-  information: {
-    color: backgroundColors.information,
-    borderColor: colors.information,
-    foregroundColor: colors.information
-  },
-  warning: {
-    color: backgroundColors.warning,
-    borderColor: colors.warning,
-    foregroundColor: colors.grayDarkest
-  },
-  danger: {
-    color: backgroundColors.danger,
-    borderColor: colors.danger,
-    foregroundColor: colors.danger
-  },
-  advisor: {
-    color: backgroundColors.advisor,
-    borderColor: colors.advisor,
-    foregroundColor: colors.grayDarkest
-  }
-};
