--- src/components/base/containers/popover/PopoverFadeTransition.js
@@ -1,13 +1,0 @@
-import React from 'react';
-
-import Fade from '../../../util/Fade';
-
-export default function PopoverFadeTransition(props) {
-  return (
-    <Fade
-      transitionClassOut="popover-transition-out"
-      transitionClassIn="popover-transition-in"
-      {...props}
-    />
-  );
-}
