--- src/components/base/containers/popover/PopoverLink.md
@@ -1,44 +1,0 @@
-```
-<p>
-  Click
-  <PopoverLink
-    popoverTitle="Popover"
-    popoverContent="This is the popover's content. Anim pariatur cliche reprehenderit, enim eiusmod high life accusamus terry richardson ad squid. 3 wolf moon officia aute, non cupidatat skateboard dolor brunch."
-  >
-    here
-  </PopoverLink> to activate the popover.
-</p>
-```
-
-```
-function PopoverForm() {
-  return (
-    <div>
-      <Textbox
-        id="firstName"
-        labelText="First name"
-      />
-      <Textbox
-        labelText="Last name"
-      />
-
-      <Button handleOnClick={() => {}}>
-        Submit
-      </Button>
-    </div>
-  );
-}
-
-const popoverContent = <PopoverForm />;
-
-<p>
-  You can even put forms inside of a <PopoverLink popoverContent={popoverContent} containsFormElement>popover</PopoverLink>. Using the <em>containsFormElement</em> attribute will lock focus to the popover, keeping tab navigation inside.
-</p>
-```
-
-```
-<p>
-  Click the question mark for some help text.
-  <PopoverLink popoverContent="Here's some help text" suppressUnderline><Icon name="question-sign" /></PopoverLink>
-</p>
-```
