+++ src/components/containers/Containers.md
@@ -1,0 +1,1 @@
+Container components handle wrapping child components with the appropriate margins and padding.
