+++ src/components/containers/tooltip/Tooltip.md
@@ -1,0 +1,11 @@
+Use a ``Tooltip`` component to display a small hover tip over content.
+
+### Basic Tooltip:
+```
+<div>Hover over <Tooltip content="This is the tooltip"><span style={{color: '#1727EE'}}>this text</span></Tooltip> to see the tooltip in action.</div>
+```
+
+### Tooltip with HTML:
+```
+<div>Hover over <Tooltip content={<span>This contains <em>some</em> <strong>HTML</strong></span>} position="left"><span style={{color: '#1727EE'}}>this text</span></Tooltip> to see the tooltip in action.</div>
+```
